-- Port to python3 completed. -- Port interface to pyqt5 completed
* I haven't tested Mayavi, since the arch version still works only on python2

* It's neccessary to polish the animations and test the app in general, there are still some glitches and there could be some minor bugs, but in general the application is working well.

Reasons:
Vtk team annouenced that vtk (version>=7.0.0) is compatible with python3 (https://blog.kitware.com/vtk-7-0-0/), therefore the mayavi team also announced support to python3. Mayabi was the only cause to not have been coded with python3 since the begining, so in order to anticipate the stables releases, I wanted to finally port all the code to the "new" python3.